+++ forge/src/main/java/top/theillusivec4/polymorph/common/integration/appliedenergistics2/PatternTermRecipesWidget.java
@@ -21,7 +21,8 @@
 
 package top.theillusivec4.polymorph.common.integration.appliedenergistics2;
 
-import appeng.menu.me.items.PatternTermMenu;
+import appeng.menu.me.items.PatternEncodingTermMenu;
+import appeng.parts.encoding.EncodingMode;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.resources.ResourceLocation;
@@ -32,10 +33,10 @@ import top.theillusivec4.polymorph.mixin.integration.appliedenergistics2.Accesso
 
 public class PatternTermRecipesWidget extends PlayerRecipesWidget {
 
-  private final PatternTermMenu container;
+  private final PatternEncodingTermMenu container;
 
   public PatternTermRecipesWidget(AbstractContainerScreen<?> containerScreen,
-                                  PatternTermMenu container, Slot outputSlot) {
+                                  PatternEncodingTermMenu container, Slot outputSlot) {
     super(containerScreen, outputSlot);
     this.container = container;
   }
@@ -53,7 +54,7 @@ public class PatternTermRecipesWidget extends PlayerRecipesWidget {
   @Override
   public void render(PoseStack pMatrixStack, int pMouseX, int pMouseY, float pRenderPartialTicks) {
 
-    if (container.isCraftingMode()) {
+    if (container.mode == EncodingMode.CRAFTING) {
       super.render(pMatrixStack, pMouseX, pMouseY, pRenderPartialTicks);
     }
   }
@@ -61,7 +62,7 @@ public class PatternTermRecipesWidget extends PlayerRecipesWidget {
   @Override
   public boolean mouseClicked(double pMouseX, double pMouseY, int pButton) {
 
-    if (!container.isCraftingMode()) {
+    if (container.mode != EncodingMode.CRAFTING) {
       return false;
     }
     return super.mouseClicked(pMouseX, pMouseY, pButton);